requirements.txt
streamlit==1.37.1
numpy==2.1.1
scikit-learn==1.5.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
openai==1.43.0
uvicorn==0.30.6

